+++ streams/AwindaStream.py
@@ -144,7 +144,7 @@ class AwindaStream(Stream):
                     data_notes=self._data_notes['awinda-imu']['timestamp'])
     self.add_stream(device_name='awinda-imu',
                     stream_name='toa_s',
-                    data_type='float32',
+                    data_type='float64',
                     sample_size=(self._num_joints),
                     sampling_rate_hz=self._sampling_rate_hz,
                     data_notes=self._data_notes['awinda-imu']['toa_s'])
